Remove file deletion on low VMAF and cleanup result extraction
Removed the code that used to delete the output file when VMAF was lower than expected in adaptiveCommand.py. This change provides more data for debugging and could prevent premature loss of valuable outputs.

Also, trimmed and sanitized the results from getHeight.py and getWidth.py scripts to avoid errors or inconsistencies in output due to unexpected leading/trailing spaces or commas. This will ensure extracted width and height data is clean and ready for use without further processing.