Add flag2 for optimal bitrate estimation
Introduced a new configuration option --flag2 to refine the encoding process for better bitrate estimation. If flag2 is enabled, the bitrate is set to the best bitrate returned by AutoBitrateLadder's method in the adaptiveAnalyser script.

This change follows research showing that estimating optimal bitrate can lead to more efficient encoding. It provides users with the flexibility to choose between different encoding strategies and also improves the accuracy of the encoding process.