Add VP8 and VP9 encoder support and use binary utilities for ffmpeg calls
This commit introduces support for VP8 and VP9 encoders and improves the function calls to ffmpeg by using binary utilities. Support for the new encoders is added in the encoder enumeration and relevant conditions are set in the VPX Encoder initiation. Function calls to ffmpeg have been modified to use the get_binary utility, ensuring correct and consistent access to the ffmpeg binary. Furthermore, the epsilon_down value in the optimisation_binary function has been adjusted for better optimization. A recurring issue with a file path in the print_and_save function has been fixed by correctly formatting the output folder path. There was an unnecessary import removed from the bitrateLadder file to clean up the code.

Signed-off-by: kingstefan26 <70776766+kingstefan26@users.noreply.github.com>